Add structures for summary counts of sparse matrix; UNFINISHED
